services/svc-alerts/requirements.txt
fastapi
uvicorn
psycopg[binary]
sqlalchemy
pydantic-settings
